NCBC-839: Implement ObserveAsync methods and overloads
Motivation
----------
Provide a means of asynchronously ensuring the durability constraints on
an operation.

Modification
------------
Add ObserveAsync to IBucket interface. Provide new asynchronous Observe
method on the KeyObserver class. Note that the synchronous Observe should
be rewritten because it mixes up asynchronous and synchronous calls. A
ticket was created to do this: NCBC-851.

Results
-------
You can now consume ObserveAsync with the async/await keywords.

Change-Id: I7054f31e0af1ae01568032a36eb68a6a7413fc80
Reviewed-on: http://review.couchbase.org/49206
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>