NCBC-840: Implement GetFromReplicaAsync methods and overloads
Motivation
----------
Provide support for "dirty" reads; provides support for reading from
replicas. Typically, a request for a key would be made using Get and if it
fails with a timed out operation, the application code could follow up
with a replica read. See docs:
http://docs.couchbase.com/developer/dev-guide-3.0/replica-read-functions.html

Modifications
-------------
The executer was updated to support asynchronous replica reads. Some unit
tests were added and an existing bug was fixed in the synchronous replica
read where the VBucket was not included in the request packet.

Results
-------
The SDK now supports synchronous and asychronous "dirty" reads from
replicated keys.

Change-Id: Id4b0a3c9d4f174348abc99f92f7e039048fccc02
Reviewed-on: http://review.couchbase.org/49270
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>