NCBC-851: Refactor KeyObserver.Observe to be synchronous
Motivation
----------
The synchronous version of Observe used some Task based constructs
internally; in certain situations this would block (for example id
Observe was called within an async block). This patch is a rewrite
of the synchronous Observe, so that it has a pure synchronous
implementation.

Modifications
-------------
The KeyObserver class was modified so that Observe would no longer use
Task based constructs. A true ObserveAsync method was added. Tests were
updated to reflect the changes made to KeyObserver. Slightly updated the
durbility logic to check for mutations after checking the replicas.

Results
-------
The synchronous Observe no longer uses Task constructs.

Change-Id: I1984cb8819ac91202c02094d59f90406ded43c98
Reviewed-on: http://review.couchbase.org/49485
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>