NCBC-847: Document.Expiry property is not honored
Motivation
----------
When doing an Insert, Replace or Upsert with a Document, the Expiry and
Cas properties should be honored. This fixes a bug where they were not
honored.

Modifications
-------------
The Insert, Replace and Upsert methods on the IBucket implementations were
modified so that the Expires properties were sent to the server. The
TimeSpanExtensions was modified with an overload which takes the
milliseconds and returns the TTL.

Results
-------
These methods will properly expire the documents with the TTL provided.

Change-Id: I5dad0ca6265d5680879edbc76046d59031541005
Reviewed-on: http://review.couchbase.org/49559
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>