NCBC-642: Use IOControl to set heartbeat settings on each TCP connection
Motivation
----------
By default, TCP keep-alives are enabled, but the client depended upon the
OS level settings for tcp_time and tcp_interval. This commit allows these
setting to be enabled on a per connection level by using IOCTL and setting
the SOL_KEEPALIVE_VALS. The default is that keep-alives are enabled and the
time and interval are set at 2hrs and 1s respectively.

Modifications
-------------
A extension method was added to the SocketExtensions class that marshals
the settings to a managed equivalent of SOL_KEEPALIVE_VALS and call
WSAIoctl via IOControl. Additionally, the
DefaultConnectionFactory.GetGeneric was modified so that this method is
called with the settings defined in the PoolsConfiguration class.

Result
------
The TCP keep-alive settings can now be set an application level
granualarity.

Change-Id: I5578c520b813cb89bb605ebd19343b24c07fbe6f
Reviewed-on: http://review.couchbase.org/49775
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>