NCBC-822: Make Transcoder and Converter pluggable
Motivation
----------
This patch makes it possible to inject custom ITypeTranscoders and
IByteConverters into the client.

Modifications
-------------
Added TranscoderElement and ConverterElements for adding configuration via
App.Config. ClientConfiguraion has two new properties: Transcoder and
Converter, which are Func factories for creating instances at runtime. The
ClusterController class was modified so that it will use the Func
factories to inject the transcoder and converters into the client.

Result
------
If the default transcoder or converter does not work for specific
use-case, it's now possible to create a custom implementation and inject
into the client.

Change-Id: I506ebdec00c9d781880c33868631473fd7e94b80
Reviewed-on: http://review.couchbase.org/49832
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>