NCBC-862: Get methods fail when stored value is null or empty
Motivation
----------
Fixes a bug where the Get operation was reading the expiry value, which is
not valid per the Memcached Binary Protocol.

Modifications
-------------
CreateExtras method was overridden in Get and the reading of the expires
was omitted.

Result
------
If a null value is stored, it can successfully be read by Get.

Change-Id: Ide373a79fa392b3a49af05a9d937da05b22c624c
Reviewed-on: http://review.couchbase.org/49840
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>