NCBC-676: Make JSON Serializer "Pluggable"
Motivation
----------
The Couchbase .NET SDK uses JSON.NET as the default serializer and it
previously was tightly coupled to the internals of the client. This commit
makes it possible to create and plug-in custom serializers using whatever
serialization dependency you chose to use.

Modifications
-------------
A new interface, ITypeSerializer, was added that defines the methods which
need to be implemented by your custom serializer. A default serializer
that uses JSON.NET was added; it is aptly named the DefaultSerializer. A
SerializerFactory was added that returns Func<ITypeSerializer> objects;
the configuration has added support for creating ITypeSerializers from a
<serializer type="" name=""/> element in the App.Config. Finally, the
constructors of all operations were refactored so that they used the value
returned by the Func<ITypeSerializer>.

Results
-------
Now, if the DefaultSerializer based off JSON.NET does not suit your fancy,
you can create a custom one (ServiceStack.Text, Jil, etc) and inject it into the client.

Change-Id: If765552821cbea6b19e14f5e61032a7ff38f1434

Conflicts:
	Src/Couchbase/CouchbaseBucket.cs
	Src/Couchbase/MemcachedBucket.cs
Change-Id: I8cfcd71e45ab8e575c21971380b71246ecd81823
Reviewed-on: http://review.couchbase.org/50209
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>