NCBC-859: Error handling and callback refactoring
Motivation
----------
Removes the need to explictly await the send request; the client now uses
a "fire and forget" strategy. All exceptions thrown will be caught and
completed by the callback and returned back to the caller.

Modifications
-------------
The IRequestExecuter implementation were refactored so that the send is
not explicitly awaited. The callback functions in CallBackFactory were
updated to handle the case where an exception was thrown and caught by the
send methods. The Server, Connection, OperationBase, and IOStrategies were updated so
that all exceptions are caught and handled correctly.

Result
------
The client now uses a "fire and forget" strategy for sending requests and
relies on all exceptions being caught and forwarded to the callback
methods for completion. The net result is a performance improvement since
the client doesn't have to wait on the send, only on the recieve side.

Change-Id: I0b5be36810c65dc5d828f27afe1ca4b3dd9d98bb
Reviewed-on: http://review.couchbase.org/50229
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>