NCBC-868: Make Connection.Dispose wait for in-flight operations to complete
Motivation
----------
When a config update occurs, in-flight operations may be canceled
abruptly; this often causes ObjectDisposedExceptions since a read may be
in process while it's dispose method is called. This commit puts a delay
on the dispose method of the connections so that it's called after either
the connection is no longer in use or after a set threshold timeout and
number of attempts.

Modifications
-------------
The IConnection interface now has properties for various states: InUse,
IsDisposed, ClosedAttempts etc. that reflect the current state of the
connection. The ConnectionPool now marks connections as being in use or
not in use when they are popped and pushed. The Dispose method of the
ConnectionPool now calls a new method CountdownToCLose(duration) method on
each connection if it's in use, otherwise the connection is disposed.

The CountdownToClose method creates a timer and checks every duration to
see if the connection is still in use, if it is the timer restarts
otherwise Dispose is called. A counter is used to determine the number of
attempts that have been made to close a connection and if it's exceeded,
Dispose will be called.

Result
------
In flight operations now have a "grace" period when a config update
triggers a reconfiguration involving disposing the connecton pool and its
connections.

Change-Id: Iaa3afa90dac0e1da07c76c0d4c7dfb3fe9947469
Reviewed-on: http://review.couchbase.org/50297
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>