NCBC-865: When cluster has no replicas GetFromReplicaAsync returns Success
Motivation
----------
This commit fixes a bug where the result returned from a replica read
would be ResponseStatus.Success and the Success field was false if a replica
read was attempted on a cluster without replicas, thus causing confusion.

Modifications
-------------
A new value for ResponseStatus was defined: NoReplicasFound. This is
returned along with Success =false if no replicas exist for the read.

Result
------
The client returns an easier to understand response state in the event
that a replica read is attempted on a cluster or server without replicas.

Change-Id: I06a0bb573cc6584b2c15505d083b20b764adfa77
Reviewed-on: http://review.couchbase.org/50337
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>