NCBC-872: Integer transcoding from 1.3.10 to 2.0.3 is incorrect
Motivation
----------
The commit fixes a bug where keys with numeric values stored by the 1.3.X
client and read with the 2.X client would be transcoded with the wrong network
byte order. Keys created with 2.X are not affected since they are use JSON
encoding and decoding and the body is always little endian.

Modifications
-------------
A flag indicating whether or not to use big endian or little endian was
add to new overloads of each method in IByteConverter; DefaultConverter
was updated with new implementations. DefaultTranscoder was updated so
that keys encoded with reserved or private flags will have their body
explicitly read/written as little endian byte arrays and conversions.
Finally, an additional parameter was added to ITypeTranscoder methods for
the operations type code, which is used to customize the endianess for
operations in which the server modifes the body: increment and decrement
for example.

Results
-------
Keys with numeric and text stored by 1.3.X can now be retrieved with 2.X.
Note that the reverse, writing a key in 2.x and reading in 1.3.X will not
work because of the flags written by 2.X are not supported by 1.3.X.
Forward compatibility only.

Change-Id: If2ab4114eee2d01f1b56aa7c5171d2dc909a1c36
Reviewed-on: http://review.couchbase.org/50434
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>