NCBC-879: Operation failure with null reference exception
Motivation
----------
During rebalance/swap/failover scenarios an endpoint may not be available
for the client to execute an operation. This commit uses a backoff
strategy to wait and try again to acquire an endpoint.

Modifications
-------------
Both memcached and couchbase IRequestExecutor's were updated so that they
use the back off strategy when when acquiring an endpoint (IServer) to send
the request.

Result
------
If the client cannot acquire an endpoint to send the query, a timeout
exception will be thrown.

Change-Id: I5cbc922c53b98a5c69785b80d8e46ec94ea888e6
Reviewed-on: http://review.couchbase.org/50579
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>