NCBC-868: Object disposed exception during rebound of rebalance 2 out
Motivation
----------
Fixes a bug where an ODE may be encountered during a failover/rebalance if
a connection is acquired between the time the pool has been disposed and
when it actually is disposed.

Modifications
-------------
Refactored the pool Dispose method so that if it is acquired, it will be
correctly disposed after the count down procedure. Additionally logic was
also added so that if the pool is disposed, it will not serve any more
connections. Also, makes sure that that SSL connection calls the base
classes dispose method which sets the inUse flag.

Results
-------
The client will wait until the connections are no longer being before
disposing them.

Change-Id: Ia3f3f6d86c3550c4f70ed0618b462b5f485c9c1f
Reviewed-on: http://review.couchbase.org/50583
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>