NCBC-809: Prevent double bootstrap of bucket
Motivation
----------
When two threads attempt to open a bucket at the same time, sometimes
a situation where the bucket is bootstraped twice has been observed,
despite the lock in ClusterController.CreateBucket method. It could be
that the lock is waited upon by the second thread, and as soon as it is
acquired it doesn't check for bucket existence but rather starts to
bootstrap it.

Modifications
-------------
ClusterController.CreateBucket now checks if the bucket has already
been created after acquiring the lock, in which case it shortcircuits
and immediately returns said bucket.
Added a test that two parallel threads opening the same bucket end up
with the same reference.

Result
------
Less room for double bootstrapping of a same bucket.

Change-Id: I4136919b3f697afb055d07a699c2c3070f589513
Reviewed-on: http://review.couchbase.org/50586
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>