NCBC-881: Ensure NRE are not thrown during retry
Motivation
----------
During rebalance/swap/failover scenarios an endpoint may not be available
for the client to execute an operation. This commit uses a backoff
strategy to wait and try again to acquire an endpoint during the retry
process.

Modifications
-------------
The CallbackFactory classes functors were updated so that they
use the back off strategy when when acquiring an endpoint (IServer) to
send the request.

Result
------
If the client cannot acquire an endpoint to send the query, a timeout
exception will be thrown.

Change-Id: Iad0a7fc5ce6dc826522ace574303b4027fc7bbed
Reviewed-on: http://review.couchbase.org/50611
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>