NCBC-882: Client does not recover from service kill/restart
Motivation
----------
If the Couchbase service crashes or is stopped, the client should
reconnect and continue to function once the service or node is
re-started.

Modifications
-------------
Added a timer to Server that starts when a dead node is detected and
checks for the connection to become available again. When it does the
client puts the node back into rotation. While the node is down, all
requests routed to it will return a ClientFailure and a ServerException
with it's IOperationResult. The message returned will be: "The current
node [HOSTNAME] is down."

Results
-------
The client will recover once a down node is re-started.

Change-Id: I7b8493db90332332ef3eda953f4aec9df8c0633f
Reviewed-on: http://review.couchbase.org/50675
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>