NCBC-885: SDK uses Nodes instead of serverList for VBucket Mapping
Motivation
----------
There is an array of Nodes elements and a serverList of hostnames in the
cluster map; the SDK uses the index of Nodes element to determine the
server to send the request to. In some cases the indexes of the nodes and
serverList are not in-synch, which causes NMVs.

This fix ensures that the index lookup is only for the hostname and then
the hostname is used to resolve the server from Nodes list.

Modifications
-------------
The ConfigContextBase, VBucket and VBucketMapper classes now use a
Dictionary<string, IServer> to hold the reference to the Server class. The
key is the hostname, when a key is mapped to a serverList index, the SDK
uses the hostname provided in the serversList as a key to look up the
Server object to use to execute the request.

Results
-------
Since VBucketMapping and the internal Servers list may change
independently, this ensures that the correct mapping is used if the index
of the serversList changes.

Change-Id: I7a5cc22121e307bdf8fc23b37d9f1aec47df57dc
Reviewed-on: http://review.couchbase.org/50806
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>