NCBC-890: When a node is dropped (e.g. server offline) OPS go to near Zero
Motivation
----------
If a endpoint or cluster suddenly goes offline, for example in the event
of a service or server crash, the client should continue to succesfully
make requests to the other still online nodes. If an operation is mapped
to a down node, it should immediately return an error result with the
reason.

Modifications
-------------
Lots of refactoring around the IServer creation; since we need to allow
authentication to happen internal to the Server class. The Server class
has a new state IsDown, which when true will make all operations routed to
it return a NodeUnavailable ResponseStatus and the exception field will
contain a NodeUnavailableException.

Result
-------
Operations that are mapped to the down node should send a failure to the client with the
IOperationResult.Exception field populated with a NodeUnavailableException
and the Status field will be NodeUnavailable. In the event of a down node, the
client will place the endpoint in a "down" state and it will start a timer and
try to reconnect at set intervals. If the goes online and the client can
connect, it will be placed back into an "active" state and continue to
process requests succesfully. Note that operations sent to a down node
will not be re-tried and the application could try a replica read if the
operation was a GET.

Change-Id: Ie3eef8b3eba35ae25f9bc3a8a67179c7729cfe38
Reviewed-on: http://review.couchbase.org/51012
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>