NCBC-893: Ensure VBucket.GetServer does not throw AOOR exceptions
Motivation
----------
In certain cases the indexes of the serverList and cluster may not be in
sync, this fix ensures that the are before accessing the value of the
index.

Modifications
-------------
Added try/catch and length checks before accessing the value at the index
provided within the serverLists and cluster arrays.

Results
-------
ArgumentOutOfRangeExceptions are handled and a null value returned if a
server cannot be found to complete the request. The client will try until
a server is found.

Change-Id: Iabc6e0af9dd4e1abe9ac1b1addede5c1ea592c01
Reviewed-on: http://review.couchbase.org/51164
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>