NCBC-896: Support translation between Hostname and IPAddress
Motivation
----------
The IPEndpointExtensions.GetServer(server) method, unlike the other
overloads, did not correctly handle the case where the passed in hostname
was not an IPAddress. This patch will correctly resolve the IPAddress of
the hostname.

Modifications
-------------
Updated IPEndpointExtensions.GetServer(server) so that if the value of
server is a hostname, the IPAddress will be resolved from it.

Results
-------
If a cluster is configured to use a hostname as opposed to a IPAddress,
the SDK will correctly resolve it.

Change-Id: I9de54a354c5f9e5952a0405696c92e596d01017d
Reviewed-on: http://review.couchbase.org/51237
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>