NCBC-901: Make IServer implementations aware of Services defined by NodesExt
Motivation
----------
In order for the SDK to take advantage of multi-dimensional scaling, the
client must maintain statefulness of which service each node is reponsible
for.

Modifications
-------------
The IService interface has added properties for which service it is
reponsible for: IsQueryNode, IsDataNode, etc. These are set by the values
in nodesExt from the cluster config. Some additional refactoring was done
around url writing and IPEndPoint and IPAddress resolution, finally unit
tests were updated.

Result
------
The internal nodes (IServer) instances are aware of the services that they
are reponsible for via configuration.

Change-Id: I80458a3a2ed21c200d4927a7da29bcb530529214
Reviewed-on: http://review.couchbase.org/51494
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>