NCBC-902: Make Server selection 'service aware'
Motivation
----------
This changeset makes the SDK 'service aware'; when a Query request is
executed only the nodes with the query service registered will execute the
query, otherwise an exception will be raised.

Modifications
-------------
The ConfigContextBase now has segragates the nodes by service into
separate lists. Accessor methods are available which will return a random
node if called. Additional properties were added to IConfigInfo to reflect
if the cluster supports a service.

Results
-------
The SDK now will send a request for an operation to the correct node if
the cluster doesn't support a service then the client will return an
exception.

Change-Id: I9b5d7df3e7b5eaa10019fac06cf0ee9cc3e4f5ae
Reviewed-on: http://review.couchbase.org/51559
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>