NCBC-904: Return error when op is attempted on node that doesn't support service
Motivation
----------
If an operation is attempted on a service that the cluster doesn't
support, throw an error with the reason.

Modifications
-------------
The Is[SERVICE]Capable flag is set when the cluster config is read and
checked when an operation is attempted. If the cluster doesn't support the
service, a ServiceNotSupportedException is thrown.

Result
------
If the cluster does not support a requested operation, the SDK will
immediatly throw a ServiceNotSupported exception.

Change-Id: I6c052d67d48c5d1dd66a3d8f338f9039738b00fd

Conflicts:
	Src/Couchbase/Configuration/CouchbaseConfigContext.cs
Change-Id: Ie11d290d42162a2a051ec8c867df6c85830b24b3
Reviewed-on: http://review.couchbase.org/51879
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>