NCBC-912: Client tries to deserialize memcached body when it is empty
Motivation
----------
If the body is empty, the SDK should bypass deserialization and the value
field should be the default value of T.

Modifications
-------------
Added a check in the DefaultSerializer to see if the length of the buffer
is 0, if it is, deserialization is bypassed.

Results
-------
The overhead of deserilization is no longer encountered when it's not
needed.

Change-Id: I636ae35784013cfbde47a212ce0a604421f5eab2
Reviewed-on: http://review.couchbase.org/51802
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>