NCBC-823: IndexOutOfRangeException when IO error occurs
Motivation
----------
If an IO error occurs while an operation is in mid-flight, the SDK would
handle the IO exception and try to read the buffer. Often the buffer would
be empty which would cause an IOORE leading to the Exception field of the
IOperationResult being overridden with the IOORE, masking the actual
error.

Modifications
-------------
Added several checks to ensure the buffer was valid before reading in
OperationBase and DefaultIOStrategy. Added a new enum value to
ResponseStatus: None, which indicates that the header has never been read.

Results
-------
The correct error should be returned back to the client. Operations that
can be re-tried such as GET and any mutation operations with a CAS value
will be retried. If the retry is not succesful, the operation may timeout.

Change-Id: I3d21dd256a77d084a3413752e2c300f372320676
Reviewed-on: http://review.couchbase.org/51890
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>