NCBC-886: Check opaque of response header against op Opaque
Motivation
----------
Adds a comparison of the opaque returned in the packet with the
operation object's opaque field to ensure that they match.

Modifications
-------------
The Read method in OperationBase now does a opaque comparison after
reading the header; if the two are not equal, a ClientFailure will be
returned and a message indicating the expected and actual opaque values.
Some of the SASL operation's didn't send an opaque, so they were modified
to do so. The Connection class now sets IsDead to true if the read request
ends up returning 0 bytes.

Results
-------
If the response of the operation does match the request, the client will
return an error.

Change-Id: I0c89c1b8d8560263507ebdc7efc8386cd9c4e1da
Reviewed-on: http://review.couchbase.org/51933
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>