NCBC-895: BufferAllocator is created per Connection
Motivation
----------
Makes the buffer for K/V IO a contiguous buffer per cluster node.

Modifications
-------------
Refactored so that BufferAllocator is now allocated at the ConnectionPool
level, instead of per connection. Changed the connection logic to use the
Offset of the SAEA assigned to itself as the offset for reading and
writing to the contiguous buffer.

Result
------
The SDK now uses the full range of indexes available and allocates the
buffer for each cluster node as a contiguous buffer.

Change-Id: If66e3eba88af5cb6ae801dbbace5cf9dd52a8e83
Reviewed-on: http://review.couchbase.org/52038
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>