NCBC-913: SDK incorrectly states down node heartbeat is 1000s instead of 1000ms
Motivation
----------
If a node unavailable, the sdk will return an error message that states
"...The SDK will continue to try to connect every 1000 seconds..". This
should be 1000 milliseconds not seconds.

Modifications
-------------
Message was modified from 1000 seconds to 1000ms in ExceptionUtil.cs

Result
------
The SDK now returns the correct error message indicating that it will try
re-connect in 1 second.

Change-Id: Ice93a0906def8bf0e22d3c8f42c45470f4875cc4
Reviewed-on: http://review.couchbase.org/52197
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>