NCBC-922: Receiving none(0xffffffff) as response status during failover
Motivation
----------
A response status of 'None' indicates that the status was never set; the
SDK should never return this value.

Modifications
-------------
Added a catch block which sets the Exception field on the operation which
will cause the GetResult and GetResultWithValue methods to return a failure
state.

Result
------
If an error occurs the proper respons status will be returned.

Change-Id: I6252b86bd9cb53a3b90ca10b6a2125ff5a66e426
Reviewed-on: http://review.couchbase.org/52288
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>