NCBC-922: Part 2 - Receiving none(0xffffffff) as response status during failover
Motivation
----------
This amends a previous commit for NCBC-922 by adding a new ResponseStatus
called TransportFailure which is returned when a SocketException is
returned indicating that their was a transport level failure on send or
recieve. In a future feature, this will make it easier to allow app to
determine which type of failures it may want to retry for example.

Modifications
-------------
ResponseStatus now has a new field called TransportFailure (0x0500) which
is returned if a SocketException is thrown and caught while sending or
receiving an operation. Catch blocks were updated so that that the
IOperationResult.Status would have this value assigned depending upon the
exception type.

Results
-------
If a SocketException is caught, the SDK will return TransportFailure; all
other client side exceptions (errors not returned in the memcached packet)
will be ClientFailure.

Change-Id: I2e03de03a2a03988ef4d74584f1b84ba938d3330
Reviewed-on: http://review.couchbase.org/52327
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>