NCBC-924: Port defaults to 0 when NodeExt is not found in cluster map
Motivation
----------
Fixes a regression in which the ports will default to 0 in pre-nodesExt
versions of Couchbase server.

Modifications
-------------
Updated NodeAdapter so that it detects the case in which nodesExt is
missing and uses the Nodes information to get the ports to use. Added unit
tests and some refactoring/formatting to existing tests.

Result
------
If the nodesExt object is missing from the cluster map, the client will
now default to the nodes array for port information.

Change-Id: I466b396272fdec816aaff5e610f6a71240f26169
Reviewed-on: http://review.couchbase.org/52347
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>