NCBC-924: part 2 - Port defaults to 0 when NodeExt is not found in cluster map
Motivation
----------
Moves the setting of the CouchbaseApiBase fields above the node
initialization section. This was omitted in the intial patch and will
cause an error.

Modification
------------
Moves the setting of the CouchbaseApiBase fields above the node
initialization section in NodeAdapter.

Result
------
If the Node is used, it will not error out when trying to parse the port
from CouchbaseApiBase.

Change-Id: I50d346d2e6925b1f100fc120d82b32d90a1dbdfa
Reviewed-on: http://review.couchbase.org/52395
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>