NCBC-932: BodyLength is zero when receive buffer tries read from it
Motivation
----------
Fixes a bug where the receive buffer length is the size of the last send.
In certain cases when the buffer is smaller than the header size, the
first receive will fail to read the entire document.

Modifications
-------------
Before the first receive is called in Connection.Send, the buffer is reset
back to the size specified in the config.

Results
-------
Documents that exceed the buffer length by less than 24 bytes will still
properly be inserted; ResponseStatus.None will not be returned.

Change-Id: I6b9f974cea267a5ba26ee48c94be81088aed79af
Reviewed-on: http://review.couchbase.org/52692
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>