NCBC-934: Ensure GetFromReplica returns an IOperationResult
Motivation
----------
In certain cases the replicas list may be empty, even though replicas are
available. If no replicas are found an IOperationResult should be
returned, not null as was happening in previous versions. Additionally,
the SDK will try to do a successful replica read until it times out or
success is returned if replicas are configured.

Modifications
-------------
Moved around the code in CouchbaseCOnfigContext to ensure that there is
not a time in which the replicas could be empty. Added logic to ensure
that IOperationResult is returned instead of null if the replicas list is
empty.

Result
------
If no replicas can be found (perhaps an inconsistent state during a
rebalance), an IOperationResult will be returned. The Result field will
indicate an OperationTimeout if it ran to timeout, NoReplicasFOund if replicas are
not configured, or success if the read was successful.

Change-Id: I06b3b7711b68b2d9ac334ad53fc2eece9f3f0aa4
Reviewed-on: http://review.couchbase.org/52766
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>