NCBC-949: Unit test cleanup and refactoring
Motivation
----------
Fixes a number of breaking unit tests; this part of a project to automate
the setup the unit test (read integration test) environment so that's it
repeatable and easy to run on a host. See NCBC-942 for details.

Modifications
-------------
A bunch of refactoring to improve the success rate of the UT's.

Results
-------
Assuming you a have a 4 node cluster with 3 replicas, SSL certs and the
following buckets defined: default, beer-sample, travel-sample,
authenticated (with 'secret' as a password) and memcached (a Memcached
bucket), you should get around 97% success rate.

Change-Id: Ifc61339804a9a7473196b5957020e35012060a79
Reviewed-on: http://review.couchbase.org/52804
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>