NCBC-941: N1QL Request Timeout is missing from QueryClient
Motivation
----------
A N1QL query request can timeout at the client level using the .NET stack and
at the server level per the REST API's timeout. The default timeout for
the server is zero which means it will never timeout; a query could hang
indefinitly. This commit adds a client side timeout.

Modifications
-------------
Adds a new field to ClientConfiguration called QueryRequestTimeout which
defaults to 70s. Updates the QueryClient to use this value to timeout the
client if a request exceeds the value.

Results
-------
The client will now timeout if a N1QL query exceeds the
QueryRequestTimeout value.

Change-Id: I69231fd13270ed678e8cac103c26c262354233e1
Reviewed-on: http://review.couchbase.org/53778
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>