NCBC-940: throw CouchbaseBootstrapException if servers list is not initialized
Motivation
----------
After receiving a cluster map and reconfigging, if the client cannot
resolve the node names resulting in the server's list being empty,
fail-fast with a CouchbaseBootstrapException.

Modifications
-------------
Add CouchbaseBootstrapException and throw after configging in the context
object if it's empty.

Results
-------
Instead of receiving an error during an operation request the client will
fail-fast while bootstrapping if it cannot properly initialize.

Change-Id: Ib179c52f043058970346f85bdfa16a53b2ad33df
Reviewed-on: http://review.couchbase.org/53906
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>