NCBC-925: .NET Client 2.1.1 leaks connections when bucket does not exist
Motivation
----------
If the .NET Couchbase Client attempts to .GetBucket() on a bucket that's
in the configuration, but does not exist, the client will throw an
AggregateException. Additionally, the client will leak connections (equal
to the min pool size) for every attempt.

Modifications
-------------
If bootstrapping fails for an IP/host the client will dispose of the
IOStrategy which will call Dipose on the internal ConnectionPool.

Results
-------
All connections are closed when boostrapping fails.

Change-Id: Iff8bf050f77a16ec2572dbab999741ced286f2ad
Reviewed-on: http://review.couchbase.org/53907
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>