NCBC-968: NRE when master node cannot be obtained during Observe
Motivation
----------
Ensure that the SDK does not throw exceptions up to the application in the
case of an operation-level error.

Modifications
-------------
KeyObserver was modified so that it will try multiple times to obtain a
node to execute the Observe operation; the CouchbaseRequestExecuter class
was modified so that uncaught exception's are handled and returned in the
IOperationResult.Exception property.

Results
-------
The client will now try several times to acquire a node; if a node cannot
be found a TimeoutException will be returned by the
OperationResult.Exception property and the Status will be
ResponseStatus.ClientError.

Change-Id: I738cf203d863f8a3bcb10c2c547537e5fff296dd
Reviewed-on: http://review.couchbase.org/54083
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>