NCBC-1023: Implement IO multiplexing/pipelining
Motivation
----------
The default IO layer uses connection pooling and a connection per thread
pattern where a thread acquires a connection pool, uses it exclusively, and
then releases it back into the pool when done. The multiplexed or
"pipelined" approach allows multiple threads to use a connection and send
requests while another dedicated thread handles the responses. The net
result is higher performance using fewer connections and less overhead on
the OS managing threads waiting for connections.

Modification
------------
This commit adds the MultipledConnection and MultiplexedIOStrategy classes
which implemenet the pattern. A new IState interface has been added to
help manage the state for an operation with async and sync
implementations. There are also changes to the ConnectionPool
class to better manage references and a new exception class for handling
the case that the server has terminated the connection. IOperation has a
new property for making it easier to track operations in async callbacks.

Result
------
The default transport layer is still DefaultIOStrategy, but by adding
adding a ioService element to your App.COnfig and by overriding the type
of the connection pool, you can switch to the multiplexed/pipelined
strategy.

Change-Id: Ia9fda9e34734b0c2f359f04862a628ebbe4ad055
Reviewed-on: http://review.couchbase.org/55063
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>