NCBC-967: Refactor bootstrapping to support MDS
Motivation
----------
With Multi-Dimensiional Scaling (MDS) a cluster node can support one or
more different services: indexer, K/V, query. This commit provides support
so that during bootstrapping that the correct node configuration is setup
correctly.

Modifications
-------------
Made changes to MemcachedConfigContext and CouchbaseConfigContext to only
create a IOStrategy for nodes that are configured for "data". Changed the
log message from ERROR to INFO for exceptions that are expected when the
client tears down the streaming config provider used for bootstrapping
non-data service nodes.

Results
-------
If the client detects that a node is node configured for "data", it will
not create any TCP connections. For memcached buckets, the internal Server
object representing a "node" will only be created for nodes configured for
"data" services.

Change-Id: Ib6fbc8c853a6882bdf3acbf7f1cda595610dc819
Reviewed-on: http://review.couchbase.org/54087
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>