NCBC-966: ScanConsistency atplus and statement plus are not supported
Motivation
----------
Couchbase Server 4.0 only supports UnBounded and RequestPlus; AtPlus and
StatementPlus are not currently supported.

Modifications
-------------
If ScanConsistency is set to AtPlus or StatementPlus a NotSupportedException
will be thrown. Also added documentation stating this and documentation
for all public members or QueryRequest class.

Result
------
AtPlus and StatementPlus are not currently supported.

Change-Id: I32c68d4a3ca6420088ee40c2357a6d87df304951
Reviewed-on: http://review.couchbase.org/54151
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>